+++ flumine/baseflumine.py
@@ -115,17 +115,18 @@ class BaseFlumine:
         for market_book in event.event:
             market_id = market_book.market_id
 
-            # check latency
-            latency = time.time() - (market_book.publish_time_epoch / 1e3)
-            if latency > 2:
-                logger.warning(
-                    "High latency between current time and MarketBook publish time",
-                    extra={
-                        "market_id": market_id,
-                        "latency": latency,
-                        "pt": market_book.publish_time,
-                    },
-                )
+            # check latency (only if marketBook is from a stream update)
+            if market_book.streaming_snap is False:
+                latency = time.time() - (market_book.publish_time_epoch / 1e3)
+                if latency > 2:
+                    logger.warning(
+                        "High latency between current time and MarketBook publish time",
+                        extra={
+                            "market_id": market_id,
+                            "latency": latency,
+                            "pt": market_book.publish_time,
+                        },
+                    )
 
             if market_book.status == "CLOSED":
                 self.handler_queue.put(events.CloseMarketEvent(market_book))